Add REPL capabilities to BattleAI
Using the readline module, we can (for now) substitute AI decisions for
user decisions for convenience so I can figure out all the different
types of messages the server will send to the client.

Add a message handler for |error|, which can result from a move choice
that was invalid.

Make Bot.addResponses immediately send response messages to the server,
because since readline's `question` method doesn't block and so any
response would be indefinitely queued until it receives some other
message from the server. This also removes the need for keeping an array
of response messages to be sent.

Since Bot is the one sending messages, it'll also need a
constructor-provided function that actually sends text to the server.
Put that in as a field which will be filled out by index.ts.