Handle |switch| messages properly
Switch just means that an active pokemon was switched out with a new
one.

Add a BattleState.toString() method for debugging purposes. Really
helpful in the long run.

Fix all pokemon being initialized as active.

Keep track of unrevealed pokemon within a team. This has yet to be fully
implemented with pokemon moves.

In Network.ts, pad or truncate input data for now so the basic neural
network still works. A true fixed input with the complete known state of
the game will be implemented in the future.

Initialize client team data in |request| messages on the first turn
only. Future |request| messages only reflect the little submessages that
come after it, only in less detail, making them ultimately unnecessary.