Allow invalid messages to finish parsing
In a private handle() method, check each argument to make sure they're
not null, then choose whether to reject the message as a whole. This
helps keep the parser in a valid state if it parses an invalid message,
and greatly simplifies parser code to not have so many `if (!x) break;`
statements.