Cover Battle tests
Add a boolean on MockBattle to track whether Battle.save() was called.

Make some message listeners async. This lets us know when all the
handlers for a certain message have finished executing.

Fix selfswitch moves not emitting a switch response.

Remove some redundant calls and an unnecessary undefined type in
RequestMove.

Expand battleTestArgs with a second requestArgs case. Also, make a note
in battleTestArgs that even indexes are reserved for the p1 PlayerID and
odd for p2.