Handle disabled status
The move Disable temporarily prevents the target from using its last
move for 4-7 turns. To allow the neural network to learn that the move
stays disabled for that long, use the reciprocal of the amount of turns
that the status was active instead of the actual amount of turns in the
neural network's input tensor, with the exception of turns=0 in which
the input would be 0. This makes a high number mean that the status is
more likely to persist on the next turn, and a low number mean that the
status is more likely to disappear on the next turn.

Expand some of the MessageParser code/tests since disable requires an
expanded message format to fully process.

Do the same as above for confusion turns. Later, this can also be
applied to a couple other temporary statuses as well as the sleep and
freeze major statuses.

Previously, the Battle object used the |request| message JSON to
determine whether a move was currently 'disabled', i.e., the move can't
be used according to the simulator, and just dumped that into the
Pokemon's VolatileStatus. Move that behavior to where it's actually
processed, which is in the Battle.getChoices() method. Additionally,
use the 'trapped' field to determine whether the AI can switch.

Ideally, the AI should not need the additional disabled/trapped info
from the |request| message JSON to determine whether it can move or
switch, and should only use it to assert that its inferences are
correct.

Move Battle's pre-event processing to right after the AI's choice is
sent to the server. This should fix cases where a turn is split up into
multiple server->client messages.