Add PossibilityClass class
This type of object is used to track a class of values and the (equal)
distribution of the probability of having either class, depending on
whether some classes have been ruled out. It essentially replaces most
of the code surrounding Pokemon.possibleHPTypes, and can be reused for
similar use cases.

Modify types dictionary to use 0-based indexes instead of booleans so it
fits with the type of dictionary the PossibilityClass constructor
requires.