Use PossibilityClass in base ability tracker
Closes #15.

In dex.abilities, make the index values start at 0 so they fit with how
PossibilityClass works.

Add a `definiteValue` field and modify `possibleValues` to include both
names and indexes. This helps with initializing the VolatileStatus when
the ability is known.

To allow VolatileStatus' overrideAbility field to not have anything,
allow it to be null so that the oneHot0 function can recognize it and
fill the oneHot array with zeros (meaning uknown value) rather than
putting a 1 in the 0th position.

Allow PossibilityClass.set to take an array, which is used to initialize
possible abilities once the species is known.