Properly encode nonexistent vs unknown
In the data array the species field was being shown as a field of 0's
when either the pokemon was nonexistent or unknown. Disambiguate that by
using a PossibilityClass object (same for items).

Add a function parameter into PossibilityClass' constructor so that it
can properly extract the uid field from the PokemonData.

Replace 1-based oneHot utility function with 0-based. The only other use
cases of the 1-based oneHot was with VolatileStatus' twoTurn and
overrideAbility field, both of which actually use a 0-based index. Fix
that so it doesn't cause any more problems.