Handle callback message
Close #5.

|callback| messages are used to provide info on an invalid choice that
was made. These can be used to gather data on any hidden factors that
are restricting the client's choices, such as being trapped by an
ability (implement later).

Change the return type of Battle.decide to an array of choices, which
should be sorted in order of preference. Keep this in a field so that if
a |callback| message is found, it will try the next preferred choice.

Make sure Network.lastChoice and lastState are up to date before the
next battleprogress message is handled using an acceptChoice method.