Add hierarchical logging
Close #35.

Add some methods to Logger that create a new Logger with some fields
modified. This allows one to create log messages with a custom
backtrace. Also separate Logger's debug and error streams.

Rewrite much of train.ts to use the Logger change, making use of
Writables to create a stream that doesn't interfere with the
ProgressBar.