Fix intermediate choices/errors desyncing training
When an intermediate choice is required (e.g. due to a self-switch move
or an ohko), the choice and corresponding prediction/state data trackers
get out of sync, causing an error on the first turn and invalid data on
subsequent turns. Handle these edge cases as well as unavailable choice
errors in TrainBattle.