Refactor Cause objects into BattleEvent fields
Close #55.

When a message suffix such as `[fatigue]` or `[from]lockedmove` is
encountered, store the value in brackets as a key of the BattleEvent
object and parse everything after it as that key's value. This allows
the PSEventHandler to better handle multiple suffixes on the same
message in the future.