Submodule Pokemon-Showdown
Parts of the server source code was needed for building certain parts of
the project, like running a local simulator for training and building a
custom dex object. To ensure consistent behavior for previous commits
and CI, start using a git submodule for this rather than cloning the
latest version of the repository.

Update references to the submodule in scripts/.