Implement tox/slp turn counter
Close #4.

Add MajorStatusCounter object to Pokemon for managing status turn
counters if applicable (slp/tox but not brn/frz/psn), which are
incremented when the status is activated.

Infer early bird ability if awake immediately after falling asleep.

Parse FromPsn suffix to indicate that damage was from poison in order to
update the toxic counter.