Use TempStatus in RoomStatus/TeamStatus fields
Close #64.

Gravity, wish, future move.

Add a `silent` field to TempStatus which can be used to automatically
call `#end()` once the duration is reached, rather than throwing on the
next tick. This is useful for wish and two-turn future moves which may
or may not receive an explicit PS event to stop their counters.

Add a `restart` parameter to `#start()` so that using wish or a future
move a second time will not reset the counter.