Infer weather item using item reference snapshot
Close #65.

Instead of keeping a reference to the Pokemon that caused a weather
condition, store a reference to its current item PossibilityClass. If
the item gets consumed or transferred, and the weather lasts 8 turns,
the Weather object will essentially be writing to a null object without
any new complex behavior having to be written.