Generalize Weather class
Close #78.

Rename Weather to ItemTempStatus and make durations and extension items
parameters of its constructor.

This is a more generalized version of Weather which should work with
other types of extendable field conditions like terrains (gen7) and
reflect/lightscreen.

Use PossibilityClass#onNarrow() to add a listener for when it's unknown
whether the source Pokemon has the status-extending item.