Handle move failure
Close #74.

Change the Pokemon#useMove() param list into a dictionary. Additionally,
distinguish between the move failing on its own or a pokemon
protecting/evading the attack so other features can be handled in the
future.

Lookahead in the |move| event handler for minor events in order to
determine whether a move failed.