Split Moveset and allow linking
Close #109.

Use a variation of the pub/sub pattern to more generally allow links to
Movesets for Transform purposes. Only #reveal() calls are propagated, so
changes to a Moveset like #replace() are not propagated in order to
support temporary Moveset-changing moves like Transform and Mimic.

Rename Pokemon#replaceMove() and #overrideMove() with sketch and mimic
respectively to better communicate their purposes.