Remove CallbackDispatcher
Close #127.

Make parser return messages instead of using MessageListener callbacks.

Add `type` fields to Message types so it's known what they are when
returned by the parser. This obsoletes the MessageListener class so
remove that from PSBot as well.

Wrap Message types in `Omit<"type">` so I don't have to comb through all
of BattleAndEventHandler.test.ts to add all those `type` fields. Since
the StateDriver (#124) is in progress and will cause the file to be
refactored anyway, use this hack until that's done.

Move Message/BattleEvent to the parser/ folder and remove the
now-obsolete CallbackDispatcher/SpecificDispatcher classes.

Replace BattleEventListener with polymorphism in PSEventHandler and
TrainEventHandler.