Encapsulate BattleState from PSEventHandler
Close #124.

This should help control what can happen to the battle state within the
game's rules, as well as reducing the amount of code in PSEventHandler
that would theoretically have to be copied if making another similar
event interpreter (e.g. another sim or cartridge).

Add a BattleDriver class which provides platform-agnostic methods for
modifying the BattleState within the game rules. It is meant to work for
multiple different game event interpreters should they be made in the
future.

Flatten PokemonID/Details/Status in some Messages/BattleEvents to make
passing data to BattleDriver a bit easier.

Move VolatileStatus#copyBoostsFrom() behavior to StateDriver since it
fits better with all the other boost mutators.