Decouple PSEventHandler from BattleDriver
Close #128.

Create DriverEvent objects (similar to BattleEvents that are received
and parsed from the PS server) that the BattleDriver can interpret. This
allows PSEventHandler methods to return DriverEvents instead of having
direct access to the BattleDriver, making PSBattle the sole owner of
both the PS-specific PSEventHandler and the platform-agnostic
BattleDriver.