Overhaul training algorithm
Use an Experience replay buffer to train the neural network while
playing training games.

Since there is no distinction between self-play and evaluation games,
simplify the logs folder structure.

During Q value calculation for learning, the algorithm has to
feedforward the network for both the current state and the next state.
The next state is for calculating the target Q value for the current
action, which is inserted into its appropriate position in the output
tensor for the current state. The current state feedforward call can be
optimized out by wrapping the network in an additional action input
which makes the network output a single Q value for the given action.