Restructure DriverEvents/BattleEvents
Close #175.

Add major/minor event distinction in BattleEvents/PSEventHandler so that
the consequences of major battle actions can be found and added under
the right DriverEvent.

Add a `consequences: AnyDriverEvent[]` field to DriverEvents in order to
discern events that cause other events to occur, e.g. a useMove with a
fail message. This simplifies much of the lookahead/lookbehind logic
previously done in PSEventHandler and makes it more simulator-agnostic.

Use some of the generalized parser types in creating the PSEventHandler
methods. Since some events require lookbehind only in PS, add a prev()
method to Iters which are now being used instead of arrays and indexes.

Add some DriverEvents to annotate useMove events: fail, miss, immune,
stall.

Add a `cause?: AnyDriverEvent` parameter to all DriverEvent handlers in
case an event needs to look behind for an event that caused it in order
to infer key info, e.g. what caused a change in weather. This also
removes the need for DriverEvent fields that contained this info, since
PSEventHandler will no longer be parsing that.

Add stalling back to the SingleTurnStatus type union, setting it
similarly to how two-turn moves are handled. Both of these effects are
explicitly set and implicitly canceled next turn, so they have to be
canceled when a move/action is used and set when their explicit event
comes up.