Use worker_threads pool to play games in parallel
Close #185.

Create a GamePool class which uses worker threads to dispatch games in
parallel.

It seemed that tfjs-node isn't meant for worker threads, since race
conditions caused several weird crashes on my system. Refactor the
training script to allow for a dedicated worker thread that handles all
the neural network prediction and training operations.

Add ESNext and WebWorkers to tsconfig.json in order to support
worker_threads as well as the @types/node package to package.json to be
able to use them in TypeScript.