Parse DriverEvent trees on StateDriver side
Close #179.

With the PSEventHandler parsing "trees" of DriverEvents via the
DriverEvent#consequence field, the limited context due to encapsulating
the BattleState made it difficult to get any meaningful relationships
between events. Move this responsibility to the StateDriver for better
control and to simplify the PSEventHandler, which wouldn't have to deal
with major/minor BattleEvents and #consequences anymore.

* The ability/trace handling for PSEventHandler is a bit weird, so move
  that to a postprocessing step within PSEventHandler#handleEvents().
  This could be moved into a separate method later if handleEvents()
  should be able to be called with fragments of a complete stream of
  DriverEvents, in which case the PSBattle should call the method before
  sending them to the BattleDriver.

To aid the StateDriver in parsing the DriverEvents, manage a stack of
sub-parsers (called "contexts") that can interpret events, add a new
sub-parser to the stack to handle a nested context, or pop themselves
off of the stack when it recognizes it's done parsing its segment.

The MoveContext was the most complicated since it had to recognize which
events were part of a move's effect in order to end the context at the
right time. Some other changes had to be made to make this easier:

* Separate nonSelfMoveCallers into separate arrays to make some
  distinctions easier.

* Remove magic coat and snatch from otherMoveCallers since they don't
  immediately call a move. Will have to test these separately later.

* Remove UseMove#targets since it can be inferred by the move and what
  happens during the move's effects.

Add StateDriver#getChoices() tests.

* Temporarily disable imprison inference on move rejection until all other
  move restrictions are properly tracked. Also disable its corresponding
  test for now.