Remove display names from dex usage
Close #203.

In dex.pokemon, replace the keys (and other occurrences of display
names) with id names to fit with the current style. In case they are
ever needed in the future, keep the display names in a separate field.

Add id name and display name fields to move dex entries. Later this can
be done for other dex groups like items and abilities if needed.