Add GPU support with command line arg
Close #204.

Add a function that imports tfjs-node or tfjs-node-gpu based on its
argument. This can be called at the main entry point when process.argv
is parsed so that all the other scripts that import tfjs will use it.

Move LearnArgs#logPath to the NetworkProcessor worker that calls it, so
that importing tfjs-node for tensorboard doesn't conflict with gpu if
selected by the worker.

Add gpu warmup step in NetworkProcessor worker.