Fix GamePoolStream/ThreadPool error recovery
Have the ThreadPool keep track of which WorkerPorts are in an errored
state, so that when the guaranteed `#givePort()` is called, it won't
throw an error comparing an old port with its replacement.

Fix ThreadPool#close() not awaiting WorkerPort#close() calls.

Remove ThreadPool#freeAllPorts() since it may conflict with pending
calls to #takePort(). Since GamePoolStream relies on this method,
have it keep track of which games it has queued to the GamePool instead.