Parallelize and shard TFRecord reading and writing
Close #210.

Move the TFRecord encoder logic to GamePort workers, with each worker
managing its own TFRecord file.

Add a thread pool class for decoding TFRecords in parallel and replace
learn()'s Dataset logic with it. Profile data indicates that this could
just be one thread to keep it out of the way of the TF worker, but
reading multiple files in parallel could sort of help with interleaving
multiple TFRecords to increase the variance in the training batches.