Simplify DriverContext stack
Close #231.

Merge DriverContext stack with the BaseContext. Instead of having
stacked contexts return a signal to defer to the BaseContext, do it
directly via inheritance or composition. This allows ContextResult to be
a `boolean | DriverContext` rather than a string union.

It seems that each of the DriverContext subclasses used some form of a
mapping from BattleEvent types to specific handlers. Refactor the base
DriverContext to more directly support this with virtual methods.

Rename BaseContext to Gen4Context, since a base context can be more
generally defined by the specific StateDriver.