Refactor BattleDriver into BattleParser function
Close #251.

Due to the nature of this refactor, some errors were introduced causing
the training script to throw for nearly every game. The next couple
commits should work on these errors.

Add chai-as-promised extension since some promise-based chai assertions
need to be used since the parser functions are now async.

Since generators have no concept of lookahead, allow the generator
parsers (SubParsers) to return a rejected lookahead event. Doing this on
the top-level should throw an error, but a SubParser calling a SubParser
should be able to handle the returned event as if it was yielded
directly to it.