Improve activateAbility/useMove handling
Close #255.

Replace SubParser's return type with an extensible interface
SubParserResult to support communication between different event
contexts (e.g. flashfire ability context informing the fire move's
context of it being blocked). This should close #252.

Add some code to the activateAbility handler to fix #253.

Add several AbilityData flags to support the above.

Remove the `on-${ctg}` qualifier from the activateAbility context's
PendingEffects builder code. This fixes #255 and simplifies the rest of
the code. Eventually this should be done for activateItem as well later.

Split the useMove event loop into separate stages to better control and
detect when implicit and ignored effects should be handled. Utilize the
extensible SubParserResult interface to control when and how each event
would be appropriate to be parsed.

Fix eventLoop not consuming all the events it can.