Handle Forewarn ability
Close #263.

Add a basePower field to MoveData specifying either the move's base
power (or 0 if variable/not applicable) or "ohko" if it's an OHKO move.

Add a warnStrongestMove flag to AbilityData to detect the Forewarn
effect rather than relying on just the name.

Add Forewarn inference code. Later this should be extended to include
the reverse case, e.g. when a pokemon switches in and Forewarn doesn't
activate, meaning either the pokemon doesn't have the ability or it
doesn't detect any attacking moves.