Add partial narrow callbacks for PossibilityClass
Close #295.

Use a modified doubly linked list for tracking subsets binded with
registered partial narrow callbacks. When a value is removed from the
possibleValues set, it also removes from the subsets it's tracking by
iterating through the list. The modification is supporting removal of
the current node while iterating in order to cleanup canceled callbacks.

Improve encapsulation by changing the PossibilityClass interface so that
it no longer exposes the possibleValues Set directly, but rather
presents it as an iterable container. Since the isSet() method is
already a substitute for Set#has(), add a #size property to complete it.