WIP: Fix training script hanging on certain errors
Change the Sender function type to return whether it successfully sent
the message it was given. This allows for setting up failsafes in the
PSBattle and the training script so that the BattleParser can properly
return and not hang in case something crashes or disconnects
unexpectedly.

Also add some more logging statements for debugging purposes.