Refactor BattleParser to use iterators
Instead of baking in iterator-like behavior into the BattleParser by
making it a generator, just make it take an iterator as an argument.

Previously SubParsers had to implement peeking/LL(1) functionality by
making most SubParsers take a `lastEvent: events.Any` argument. By using
a custom pair of mutually exclusive AsyncIterators this behavior can be
implemented directly. Additionally, exceptions had to be propagated
across both iterators as well as to the final Promise to finish the
BattleParser. This makes the implementation and error handling a bit
more complicated but the tests and test runs verify that it all works
well.