Overhaul model
Close #323.

Use shared weights and multiple inputs to remove dependence on input
order for pokemon/moves.

Rewrite state encoders in order to reflect this change.

Add some custom layers to deal with a few specific ops that tfjs-layers
didn't support.

The new model has a total of 69,435 parameters, a 660x decrease compared
to the original 45,878,282 due to the maximal amount of weight sharing
being applied. Some initial testing shows a dramatic reduction in GPU
usage, so some tweaking/upsizing of the model or batch size may be
needed to improve performance.

The new model takes 31 different inputs (since slicing would tightly
couple with the encoder which can be unstable), totaling up to 42,146
floats, similar to the original 44,792. This was due to some
simplifications made in the encoder to move the VolatileStatus'
lastMove, disabled, and encore encodings into the encoding for the Move
that they mention.