Prevent fainted team members from affecting model
Add an additional input tensor used to mask out dead team members so
they don't get included in the team pokemon features aggregate layer in
the model.

Since tf.layers.repeatVector and tf.layers.dot don't work well with
having an extra dimension for the team, add a custom layer which does
the broadcasting logic in the way we want in order to apply the mask.