Combine replay buffer implementations
Combine ReplayBuffer and TrajectoryReplayBuffer into a new class that
doesn't make assumptions on the structure of the stored examples and
tries to batch them the best it can.

Also fix a bug in agent.select_action() where the new hidden states were
being converted into tuples before being stored in the trajectory. This
never affected the original TrajectoryReplayBuffer which forced them
back to lists when batching, but it's now required here to ensure
consistency.