Fix break during for-await stream causing error
Possible fix for #365.

This should fix any cryptic "AbortError: The operation was aborted"
errors when running on Node.js versions above v16. It seems that a few
breaking changes were made to some stream behaviors in v17(?) that
caused for-await-of loop interruptions to be treated as errors by the
stream's implicit destroyer. Explicitly calling end() here assures
better handling of those interruptions.